[ECHO] changed THE WHOLE function. Needs testing.

+++ src/includes/hardbs.h
@@ -2,9 +2,9 @@
 #define HARDBS_H
 int isforbidden(char c);
 
-int back_U(char *arg, char *printfin, int index, size_t *i);
-int back_u(char *arg, char *printfin, int index, size_t *i);
-int back_x(char *arg, char *printfin, int index, size_t *i);
-int back_zero(char *arg, char *printfin, int index, size_t *i);
+int back_U(char *arg, char *printfin, int index, int *i);
+int back_u(char *arg, char *printfin, int index, int *i);
+int back_x(char *arg, char *printfin, int index, int *i);
+int back_zero(char *arg, char *printfin, int index, int *i);
 
 #endif /* HARDBS_H */
